testProject/requirements.txt
#Python 3.6.8
aniso8601==7.0.0
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.0.2
django-cors-headers==2.4.0
django-environ==0.4.5
django-graphql-jwt==0.3.0
djangorestframework==3.11.0
djangorestframework-expiring-authtoken==0.1.4
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.1
idna==2.8
pkg-resources==0.0.0
promise==2.3
psycopg2-binary==2.7.7
PyJWT==1.7.1
pysolr==3.8.1
pytz==2019.3
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7